refactor: 생성자 DI 테스트 패키지 정리
