+++ app/src/test/java/woowacourse/shopping/di/DefaultFakeRepository.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.di
+
+class DefaultFakeRepository : FakeRepository
