+++ app/src/test/java/woowacourse/shopping/di/FakeRepository.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.di
+
+interface FakeRepository
