+++ app/src/test/java/woowacourse/shopping/di/constructordi/FakeConstructorDIActivity.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.di.constructordi
+
+import woowacourse.shopping.di.DIActivity
+
+class FakeConstructorDIActivity : DIActivity() {
+    lateinit var viewModel: FakeConstructorDIViewModel
+}
