+++ app/src/test/java/woowacourse/shopping/di/constructordi/FakeConstructorDIApplication.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.di.constructordi
+
+import woowacourse.shopping.di.DIApplication
+import woowacourse.shopping.di.Injector
+
+class FakeConstructorDIApplication : DIApplication() {
+    override fun inject() {
+        Injector(container).inject(FakeConstructorDIModule)
+    }
+}
