+++ app/src/test/java/woowacourse/shopping/di/constructordi/FakeConstructorDIViewModel.kt
@@ -1,0 +1,6 @@
+package woowacourse.shopping.di.constructordi
+
+import androidx.lifecycle.ViewModel
+import woowacourse.shopping.di.FakeRepository
+
+class FakeConstructorDIViewModel(val fakeRepository: FakeRepository) : ViewModel()
